refactor: extract admin task projection
