+++ crates/dcc-mcp-gateway/src/gateway/admin/mod.rs
@@ -25,7 +25,7 @@
 //!
 //! The runtime UI is a single inline HTML string (`admin/html.rs`). The
 //! `dcc-mcp-gateway-admin` crate owns the trace domain, compact, link,
-//! issue-report, statistics, analytics, governance, artifact, activity, and traffic projections,
+//! issue-report, statistics, analytics, governance, artifact, activity, task-outcome, and traffic projections,
 //! Vite/npm build boundary, and embedded asset; this module owns the
 //! gateway-specific data-source/HTTP adapters and API handlers.
 //!
@@ -35,7 +35,7 @@
 //! admin/
 //! ├── domain/       # pure types, no I/O (trace types)
 //! ├── application/  # orchestration/handler routing
-//! ├── infra/        # SQLite reads, log reads, activity projection, integration config
+//! ├── infra/        # SQLite reads, log reads, projection adapters, integration config
 //! ├── html.rs       # standalone asset module (feature-gated)
 //! └── mod.rs        # re-exports only
 //! ```
@@ -64,7 +64,7 @@ pub mod domain;
 /// Application layer: handler routing and orchestration.
 pub mod application;
 
-/// Infrastructure layer: SQLite, logs, activity projection, integration config.
+/// Infrastructure layer: SQLite, logs, projection adapters, integration config.
 pub mod infra;
 
 // ── Backward-compat shim modules (content moved to domain/application/infra) ──
